requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Babel==2.14.0
beautifulsoup4==4.12.3
boto3==1.34.15
botocore==1.34.15
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-countries==7.5.1
django-dbbackup==4.1.0
django-easy-audit==1.3.5
django-import-export==3.3.7
django-jazzmin==2.6.0
django-storages==1.14.2
docxcompose==1.4.0
docxtpl==0.16.8
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.5
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==10.0.1
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyotp==2.9.0
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.18
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
